### 2020-08-04 #2546 @whedon
[PRE REVIEW]: s4rdm3x: A Tool Suite to Explore Code to Architecture Mapping Techniques
**Submitting author:** @tobias-dv-lnu (<a href="http://orcid.org/0000-0003-1154-5308">Tobias Olsson</a>)
**Repository:** <a href="https://github.com/tobias-dv-lnu/s4rdm3x" target ="_blank">https://github.com/tobias-dv-lnu/s4rdm3x</a>
**Version:** 1.2
