### 2020-08-04 #2546 @gkthiruvathukal
@danielskatz Yes,  I can take this one on. This might be the last I can take on for a couple of weeks as I have to record a bunch of lectures for my upcoming classes this term.  I won't go into OOO mode as I plan to see all submissions through, but I am going to need to pause on new ones just for a short time.  By the way, I am assuming you meant to invite me as an editor, right?
