### 2020-09-14 #2546 @kinow
I think a friend of mine was doing some bytecode analysis, but something much simpler than `s4rdm3x` (I think it was simply looking for patterns in the bytecode or in the abstract syntax tree). `s4rdm3x` looks interesting!

Would be happy to review this one @gkthiruvathukal
