### 2020-10-28 tobias-dv-lnu/s4rdm3x#6 @kinow
s4rdm3x_1.2.zip missing from GH release?
Hi @tobias-dv-lnu ,

I'm following the installation instructions, and it looks like the latest release on GH is missing a ZIP that's present in the other two? I'm downloading a src file, but not sure if it will have the jars I need to keep following the docs?

Related to <https://github.com/openjournals/joss-reviews/issues/2791>

Thanks
Bruno
