### 2020-10-29 tobias-dv-lnu/s4rdm3x#6 @tobias-dv-lnu
Yes, release 1.2 was a source-only release, basically adding the gradle stuff. Zip with jars from 1.1 should be fine. I will add the 1.1 release zip to release 1.2 to make it consistent with the documentation.

Sorry for the confusion.

/T
