### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @kinow
Failed to load `experiment.xml` with Ubuntu and `v3xt.jar`
From https://github.com/openjournals/joss-reviews/issues/2791

I'm getting the following when trying to load the sample experiment.

```
$ ~/Development/java/jdk-11.0.2/bin/java -jar v3xt.jar
Exception when attempting to load/save:/home/kinow/Downloads/1.1/experiment.xml
java.io.IOException: Could not read System File: data\systems\JabRef\3.7\jabref-3_7.sysmdl
java.io.IOException: Could not read System File: data\systems\JabRef\3.7\jabref-3_7.sysmdl
	at se.lnu.siq.s4rdm3x.experiments.system.FileBased.<init>(FileBased.java:32)
	at se.lnu.siq.s4rdm3x.experiments.ExperimentXMLPersistence.elementToRunner(ExperimentXMLPersistence.java:149)
	at se.lnu.siq.s4rdm3x.experiments.ExperimentXMLPersistence.loadExperimentRunners(ExperimentXMLPersistence.java:74)
	at experimenting.ExperimentsView.doSaveButtons(ExperimentsView.java:225)
	at experimenting.ExperimentsView.doView(ExperimentsView.java:116)
	at Main.mainLoop(Main.java:181)
	at Main.lambda$run$0(Main.java:144)
	at uno.glfw.GlfwWindow.loop(GlfwWindow.kt:654)
	at Main.run(Main.java:135)
	at Main.main(Main.java:54)

```

Any idea what I got wrong?
