### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @kinow
I am using the release 1.2 (even though the directory says 1.1). And the file that it's failing to load appears to exist in the expected location. Might be some issue with how the file is loaded, or maybe it works on Mac/Win but not Linux, or maybe it's something else with my environment?

![image](https://user-images.githubusercontent.com/304786/97838622-7783b000-1d45-11eb-8a82-4ff8e5675ffd.png)
