### 2020-11-02 tobias-dv-lnu/s4rdm3x#11 @kinow
Add a CONTRIBUTING.md file or instructions for contributors?
Hi

From https://github.com/openjournals/joss-reviews/issues/2791

Just wondering if there are any plans to document how others can contribute to `s4rdm3x`? Related to the following topic in the review process

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Or maybe there's something already in the docs that I've missed?

Thanks!
