### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
A likely error is that the example experiment file uses windows type file delimiters (as it was saved on my windows dev machine).

The line "<system file="data\systems\JabRef\3.7\jabref-3_7.sysmdl" name="JabRef_37_1"/>" should be changed to:

<system file="data/systems/JabRef/3.7/jabref-3_7.sysmdl" name="JabRef_37_1"/>

If you can confirm that this is the problem, I can make some code changes to save the file path in a system-neutral way when the experiments are saved.
