### 2020-11-02 tobias-dv-lnu/s4rdm3x#11 @tobias-dv-lnu
There is a small section at the end of the docs: https://tobias-dv-lnu.github.io/s4rdm3x/

The main way of continuing development is to write new mappers: the main steps to do so is documented in: https://tobias-dv-lnu.github.io/s4rdm3x/add_new_mapper

Maybe this should be made clearer in some way...
