### 2020-11-02 tobias-dv-lnu/s4rdm3x#11 @kinow
Sorry, I missed that part @tobias-dv-lnu . I think a simple short paragraph or link to the section on how to contribute/write a custom mapper could be useful, or what do you expect from the contribution (e.g. include tests? add javadocs to important methods/code added, etc?).

But good enough for the paper review :+1: plus I consider the vast amount of tests and the example experiment provided good ways to help users how to contribute too (i.e. by example). Feel free to close it if you'd like :+1:
