### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @kinow
@tobias-dv-lnu I think the code on `master` might be working. When I loaded the `ExperimentXMLPersistence.java` class in Eclipse, I realized it was taking care of replacing some characters by the `File.fileSeparator` as in [here](https://github.com/tobias-dv-lnu/s4rdm3x/blob/ab286f8a240f7049362e3a061bd305d29fca1fc3/src/main/java/se/lnu/siq/s4rdm3x/experiments/ExperimentXMLPersistence.java#L150).

I debugged the jar from Eclipse (`-Xdebug...` specifying the remote port) after checking from the tag, and confirmed the path is indeed using the wrong separator.

![image](https://user-images.githubusercontent.com/304786/97847873-267bb800-1d55-11eb-9fa8-31ed7161dd30.png)

I don't know how to generate the two jars with Gradle @tobias-dv-lnu , so I'm unable to modify the source code and confirm it's really working from `master`. Maybe with a new release we can assess it again?

Thanks
B
