### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
Yes, you are correct.

It seems there can be an attribute "file_separator" in the XML that should specify the separator used when the file was saved, it should then be replaced with the platform-specific separator `File.separator`

The experiment XML in the release seems to be missing this attribute, however, as per the code it should default to the more odd windows specific separator \. I.e. \ should be replaced with / on Linux.

I will investigate and create a new release asap.
