### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
I have confirmed the code working in cmdexrunner (it uses the same code but is command line based) on an ubuntu-machine. I will create a new release.
