### 2020-11-02 tobias-dv-lnu/s4rdm3x#10 @kinow
I started `v3xt.jar`, then clicked on the "Load" button.

```bash
$ ~/Development/java/jdk-15/bin/java -jar v3xt.jar
Could not add system model: data/systems/argouml/argouml.sysmdl - referenced jars do not exist
Could not add system model: data/systems/ant/ant.sysmdl - referenced jars do not exist
Could not add system model: data/systems/JabRef/3.7/jabref-3_7.sysmdl - referenced jars do not exist
Could not add system model: data/systems/sweethome3d/sweethome3d.sysmdl - referenced jars do not exist
Could not add system model: data/systems/ProM6.9/ProM_6_9.sysmdl - referenced jars do not exist
Could not add system model: data/systems/teammates/teammates.sysmdl - referenced jars do not exist
Could not add system model: data/systems/lucene/lucene.sysmdl - referenced jars do not exist
```

This is a different box, but also with Ubuntu latest LTS (20.04).
