### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
Yes, that would be normal as you also need to download the systems to analyze (I cannot include them in the distribution due to licensing).

I suggest you download JabRef 3.7 jar and place it in the same directory as the jabref-3_7.sysymdl

You can find it here: https://github.com/JabRef/jabref/releases/tag/v3.7
