### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @kinow
Hi @tobias-dv-lnu

I think I'm understanding more how `s4rdm3x` works. From what I understood, there are several projects in the `systems` folder. I thought they were dependencies for the experiments, but looks like they are projects which you can run the experiments against?

Which means if I wanted to test Apache Commons Lang, I would need to create a folder `data/systems/commons-lang`, and create a `commons-lang.sysmdl` file with the details of the project architecture.

Is that correct?

>This is mentioned in the documentation under Running/Data Systems. A better error message would also be good.

I think we could just move or link the sections. I am following the docs from the top to the bottom. Looking one section at a time. I saw in one section that I could run the `v3xt.jar`, and immediately started playing with it.

Maybe in the end of that section, we could mention that users are expected to read the next section before running/loading the existing experiments. Something like "In order to run the experiments you will have to first download at least one of the projects used. We will see how this can be done in the next section etc etc", or something along these lines?

WDYT?
