### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
Hey!

_I think I'm understanding more how s4rdm3x works. From what I understood, there are several projects in the systems folder. I thought they were dependencies for the experiments, but looks like they are projects which you can run the experiments against?_

Exactly!

_Which means if I wanted to test Apache Commons Lang, I would need to create a folder data/systems/commons-lang, and create a commons-lang.sysmdl file with the details of the project architecture.

Is that correct?_

Correct! The .sysmdl are quite simple text files, so you can create one in a text editor. The problem is, of course, that you need a mapping for the source code to the architectural modules these are often created by experts or at least need a large effort and deep knowledge of the system. From what I know I have the largest number of such models included in this repo.

You are correct about the documentation, I guess most would read from top to bottom and then start to play with things. I will make some changes.

Thanks!
