### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @kinow
I loaded the experiment.xml again, after putting the jar in the folder of JabRef as you said @tobias-dv-lnu . It printed the errors for the other applications, but I am ignoring them for now.

I also pressed "Run Experiment". There was no clear indication that the experiment was running successfully for JabRef (or not one that was clear to me?). Is there somewhere I could accompany via logs or GUI what's happening?

After a while I stopped with the button Stop Experiment.

And finally launched the `cmdexrunner.jar`:

```bash
$ ~/Development/java/jdk-11.0.2/bin/java -jar cmdexrunner.jar -experiment experiment.xml -outDir bruno
Running 1 threads. Saving data in: bruno/JabRef_37_1_0
Running experiments:
All experiment threads Started!
0, WARNING: core mtj jar files are not available as resources to this classloader (jdk.internal.loader.ClassLoaders$AppClassLoader@799f7e29)
```

Are these the correct steps @tobias-dv-lnu ? I am trying to get a quick overview of s4rdm3x, and then read the code and paper after I understood the basic operation of the tool.

The `cmdexrunner.jar` produced an CSV with some metrics that I plan to take a closer look later. Screenshot of the CSV below.

![image](https://user-images.githubusercontent.com/304786/97960814-001e5100-1e17-11eb-86b0-f5a012367abd.png)

My intention is to run s4rdm3x against Apache Commons Lang. I remember using a tool that appears to be similar (JMack? MackJ?). I remember I could define the packages/modules of my code, and how they should relate to each other. That tool would then run a static analysis and give me the report saying whether there were architectural issues or not. Though I still need to read more about the theory behind `s4rdm3x` before I can compare both…
