### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
You should get some results from the experiments in the plots windows (however, as default they may be hidden or very small). Make sure JabRef is checked before you start the experiment (it should be but you never know). And then check/enlarge the plot windows. You should see some results after a few seconds.

![image](https://user-images.githubusercontent.com/3469312/97962009-98cfb500-1db4-11eb-99f2-aea1fe32c11b.png)
