### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
Interesting case with Apache, let me know if/how I can help.

In the research community, the main tools used for static architectural conformance checking are JabRef (the sysmdl format is based on an older version of the JabRef format) and Hussact, but there are also tools/experiments that analyze source code and try to find problems or abstractions. For some reason, JMack/MackJ sounds familiar but I cannot recall exactly where I heard it.

I'm mainly working on trying to automate the mapping of source code to the architectural modules as this is a prerequisite to doing static architectural conformance checking. s4rdm3x is about testing approaches to automating such mappings and comparing with a human made mapping (that is why a ground truth mapping is needed).
