### 2020-11-03 tobias-dv-lnu/s4rdm3x#10 @tobias-dv-lnu
FYI: I moved the Data systems part of the documentation to be a sub-section to Installation. I.e. first you download the program, then you get instructions on the data systems needed, and finally instructions on how to run.

Much better imho :)
