### 2020-11-05 tobias-dv-lnu/s4rdm3x#10 @kinow
Installation instructions working @tobias-dv-lnu . Successfully ran my first experiment with `s4rdm3x`.

![image](https://user-images.githubusercontent.com/304786/98213333-caf63800-1fa9-11eb-9cf3-e9a21a90134a.png)
