### 2020-11-05 #2791 @kinow
My review is complete @gkthiruvathukal :heavy_check_mark:

Even managed to test `s4rdm3x` with Apache Commons Imaging. I was going to use Apache Commons Lang, but I am currently working on Imaging. I didn't add all the relationships between packages, and the tool correctly highlighted the failed clusterings.

![image](https://user-images.githubusercontent.com/304786/98224880-e61c7400-1fb8-11eb-83f5-c84cbbd89cec.png)

Didn't get fully familiar with the rest of the interface and metrics, but looks really useful for developers inspecting code bases, and getting familiar with new architectures.

I used IntelliJ's diagram builder to remind me of some of the relationships, but in the past I used Stan4J that would be a good tool to help users to create `sysmdl` files.

<details><summary>commons-imaging.sysmdl</summary>
# name
Apache Commons Imaging

# jar
commons-imaging-1.0-alpha2.jar

# root-packages
org/apache/commons/imaging/

# metrics file
#commons-imaging-file-metrics.csv

# modules
base
color
common
bmp
dcx
gif
icns
ico
jpeg
pcx
png
pnm
psd
rgbe
tiff
wbmp
xbm
xpm
icc
internal
palette

# <high_level_module> <regular_expression>
base org.apache.commons.imaging.color.[^.]*
color org.apache.commons.imaging.color.*
common org.apache.commons.imaging.common.*
bmp org.apache.commons.imaging.formats.bmp.*
dcx org.apache.commons.imaging.formats.dcx.*
gif org.apache.commons.imaging.formats.gif.*
icns org.apache.commons.imaging.formats.icns.*
ico org.apache.commons.imaging.formats.ico.*
jpeg org.apache.commons.imaging.formats.jpeg.*
pcx org.apache.commons.imaging.formats.pcx.*
png org.apache.commons.imaging.formats.png.*
pnm org.apache.commons.imaging.formats.pnm.*
psd org.apache.commons.imaging.formats.psd.*
rgbe org.apache.commons.imaging.formats.rgbe.*
tiff org.apache.commons.imaging.formats.tiff.*
wbmp org.apache.commons.imaging.formats.wbmp.*
xbm org.apache.commons.imaging.formats.xbm.*
xpm org.apache.commons.imaging.formats.xpm.*
icc org.apache.commons.imaging.formats.icc.*
internal org.apache.commons.imaging.formats.internal.*
palette org.apache.commons.imaging.formats.palette.*

# <source_module> <target_module>

base icc
base common
common base
common internal
icc common
jpeg common
palette base
palette internal
color internal
