### 2020-11-05 #2791 @kinow
> Awesome! @kinow: Do you consider the sysmdl for commons imaging to be "correct"? If so would can I include it in the repo as an available system (always on the lookout for more data :) )

Not yet, but I will open a draft pull request. We can discuss how to improve it. I didn't understand how to use all the options and settings, and for the `experiment.xml` I simply used the same one from JebRef modifying the system loaded.

I've created a draft PR: https://github.com/tobias-dv-lnu/s4rdm3x/pull/14
