### 2021-01-28 #2791 @tobias-dv-lnu
release version tag: v1.3_archive
URL: https://github.com/tobias-dv-lnu/s4rdm3x/tree/v1.3_archive
DOI: 10.5281/zenodo.4475664
URL: https://zenodo.org/record/4475664#.YBKnmjpKguU
