### 2021-02-05 #2791 @kyleniemeyer
@tobias-dv-lnu, it looks like DOIs may be available but missing for a number of your references. Can you please check these, and add if appropriate? I will then be able to accept the submission. Thanks!
