### 2021-02-06 #2791 @tobias-dv-lnu
@kyleniemeyer Thanks! Just added the missing DOIs in the paper.bib file (whedon had made the correct guesses on the DOIs), everything should be ok now I hope :)
